memory: jz47xx-nemc: driver for the NEMC on JZ47xx SoCs
Add a driver for the NAND/External Memory Controller (NEMC) on JZ4760
and later SoCs.

The primary function of this driver is to configure parameters, such
as timings, for external memory devices using data supplied in the
device tree. Devices connected to the NEMC are represented in the DT
as children of the NEMC node, the driver uses optional properties
specified in these child nodes to configure the parameters of each
bank.

[paul.burton@imgtec.com: fixup merge conflicts]

Signed-off-by: Alex Smith <alex.smith@imgtec.com>