CHROMIUM: MIPS: GIC: Use local interrupts for timer
Instead of using GIC interrupt 0 for the timer (which was not even
handled correctly by the GIC irqchip code and could conflict with an
actual external interrupt), use the designated local interrupt for
the GIC timer.

Also, since the timer is a per-CPU interrupt, initialize it with
setup_percpu_irq() and enable it with enable_percpu_irq() instead
of using direct register writes.

BUG=chrome-os-partner:30840
TEST=Build and boot on Malta and Danube; GIC timer now works on Danube.

Change-Id: Id3cd88b0d3a9387cbb4120149d3f5a94aeb57141
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>