CHORMIUM: MIPS: Malta: Map GIC local interrupts
Now that the GIC driver properly supports local interrupts, extend
the static interrupt mapping to include the GIC timer and watchdog
and fix up the GIC interrupt setup and handling so that the local
interrupts are properly handled.  Note that ipi_map is also renamed
to gic_irq_map since it is now also used to track mapping of non-IPI
interrupts to CPUs.

BUG=chrome-os-partner:30840
TEST=Build and boot on Malta and Danube; GIC timer now works on Malta.

Change-Id: Ie7b58f6da643e0da514b1b25b6c224691251c0be
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>